soldajustica/requirements.txt
django<2
django-suit==0.2.13
django-suit-redactor==0.0.2